refactor(#243): trash·user 도메인을 FSD 구조에 맞게 정리
