+++ src/entities/trash/api/trash.api.ts
