+++ src/entities/trash/types/trash.types.ts
