+++ src/features/trash/hooks/useDeleteTrashMutation/useDeleteTrashMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-import { deleteTrash } from "@/features/trash/api/trash.api";
-import { TrashActionParam } from "@/features/trash/types/trash.types";
+import { deleteTrash } from "@/entities/trash/api/trash.api";
+import { TrashActionParam } from "@/entities/trash/types/trash.types";
 import { useToast } from "@/shared/hooks/useToast";
 import { ApiError } from "@/shared/lib/api/types";
