+++ src/features/trash/hooks/useRestoreTrashMutation/useRestoreTrashMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-import { restoreTrash } from "@/features/trash/api/trash.api";
-import { TrashActionParam } from "@/features/trash/types/trash.types";
+import { restoreTrash } from "@/entities/trash/api/trash.api";
+import { TrashActionParam } from "@/entities/trash/types/trash.types";
 import { useToast } from "@/shared/hooks/useToast";
 import { ApiError } from "@/shared/lib/api/types";
