+++ src/widgets/trash/PersonalTrash/PersonalTrash.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { trashQueries } from "@/features/trash/trash.queryKey";
+import { trashQueries } from "@/entities/trash/query/trash.queryKey";
 import { useInfiniteScroll } from "@/shared/hooks/useInfiniteScroll";
 import TrashEmpty from "@/widgets/trash/TrashEmpty";
 import TrashList from "@/widgets/trash/TrashList";
