+++ src/widgets/trash/TeamTrash/TeamTrash.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { trashQueries } from "@/features/trash/trash.queryKey";
+import { trashQueries } from "@/entities/trash/query/trash.queryKey";
 import { useInfiniteScroll } from "@/shared/hooks/useInfiniteScroll";
 import TrashEmpty from "@/widgets/trash/TrashEmpty";
 import TrashList from "@/widgets/trash/TrashList";
