+++ src/widgets/trash/TrashList/TrashList.tsx
@@ -1,9 +1,9 @@
 "use client";
 import React, { useState } from "react";
 
+import { TrashItemData } from "@/entities/trash/types/trash.types";
 import { useDeleteTrashMutation } from "@/features/trash/hooks/useDeleteTrashMutation";
 import { useRestoreTrashMutation } from "@/features/trash/hooks/useRestoreTrashMutation";
-import { TrashItemData } from "@/features/trash/types/trash.types";
 import SoftButton from "@/shared/ui/Button/SoftButton";
 
 import TrashItem from "../TrashItem";
